// src/app/shared/models/servicetype.ts
export class ServiceType {
    id: number;
    name: string;
    displayName: string;
    description: string;
    iconUrl: string;
    colorCode: string;
    isActive: boolean;
    isDelete: boolean;
    createdDate: Date;
    modifiedDate: Date;
    createdBy: number;
    modifiedBy: number;
    constructor() { }
}